refactor(scalars): refactor storybook structure
