+++ packages/design-system/src/scalars/components/examples/enum-field-example/enum-field-example.tsx
@@ -1,18 +1,8 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import EnumFieldExample from "./enum-field-example.js";
+import { Button } from "#powerhouse";
+import { EnumField } from "../../enum-field/index.js";
+import { Form } from "../../form/index.js";
 
-const meta = {
-  title: "Document Engineering/Docs/Examples/Enum Field Example",
-  component: EnumFieldExample,
-  tags: ["autodocs"],
-  parameters: {
-    layout: "centered",
-    docs: {
-      source: {
-        language: "tsx",
-        format: true,
-        code: `
-function EnumFieldExample() {
+const EnumFieldExample = () => {
   const onSubmit = async (data: any) => {
     // simulate a network request
     await new Promise((resolve) => setTimeout(resolve, 1000));
@@ -35,9 +25,7 @@ function EnumFieldExample() {
   ];
 
   return (
-    <Form
-      onSubmit={onSubmit}
-    >
+    <Form onSubmit={onSubmit}>
       {({ formState: { isSubmitting } }) => (
         <div className="flex w-[400px] flex-col gap-4">
           <EnumField
@@ -66,16 +54,6 @@ function EnumFieldExample() {
       )}
     </Form>
   );
-}`,
-      },
-    },
-  },
-} satisfies Meta<typeof EnumFieldExample>;
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  args: {},
 };
+
+export default EnumFieldExample;
