+++ packages/design-system/src/scalars/components/examples/multiple-fields-with-complex-layout/multiple-fields-with-complex-layout.stories.tsx
@@ -3,7 +3,7 @@ import MultipleFieldsWithComplexLayout from "./multiple-fields-with-complex-layo
 
 const meta = {
   title:
-    "Document Engineering/Docs/Examples/Multiple Fields With Complex Layout",
+    "Document Engineering/Scalars/Examples/Multiple Fields With Complex Layout",
   component: MultipleFieldsWithComplexLayout,
   tags: ["autodocs"],
   parameters: {
