+++ packages/design-system/src/scalars/components/examples/submit-changes-only/submit-changes-only.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 import SubmitChangesOnly, { DOCS_CODE } from "./submit-changes-only.js";
 
 const meta = {
-  title: "Document Engineering/Docs/Examples/Submit Changes Only",
+  title: "Document Engineering/Scalars/Examples/Submit Changes Only",
   component: SubmitChangesOnly,
   tags: ["autodocs"],
   parameters: {
