+++ packages/design-system/src/scalars/components/examples/submit-changes-only/submit-changes-only.tsx
@@ -1,5 +1,6 @@
 import { Button } from "#powerhouse";
-import { Form, StringField } from "../../../components/index.js";
+import { Form } from "../../form/index.js";
+import { StringField } from "../../string-field/index.js";
 
 const SubmitChangesOnly = () => {
   return (
