+++ packages/design-system/src/scalars/components/form/forms.mdx
@@ -1,10 +1,10 @@
 import { Meta, Source, Canvas, Controls } from "@storybook/blocks";
 
-import * as ResetButtonStories from "./examples/reset-button/reset-button.stories";
-import * as ResetOnSuccessfulSubmitStories from "./examples/reset-on-successful-submit/reset-on-successful-submit.stories";
-import * as SubmitChangesOnly from "./examples/submit-changes-only/submit-changes-only.stories";
+import * as ResetButtonStories from "../examples/reset-button/reset-button.stories";
+import * as ResetOnSuccessfulSubmitStories from "../examples/reset-on-successful-submit/reset-on-successful-submit.stories";
+import * as SubmitChangesOnly from "../examples/submit-changes-only/submit-changes-only.stories";
 
-<Meta title="Document Engineering/Docs/Forms" />
+<Meta title="Document Engineering/Scalars/Forms" />
 
 # Introduction
