+++ packages/design-system/src/scalars/components/fragments/id-autocomplete-field/id-autocomplete-input-container.tsx
@@ -4,7 +4,7 @@ import { cn } from "#scalars";
 import { Command as CommandPrimitive } from "cmdk";
 import React, { useState } from "react";
 import { useMediaQuery } from "usehooks-ts";
-import { Input } from "../../../../ui/components/data-entry/index.js";
+import { Input } from "../../../../ui/components/index.js";
 import {
   Tooltip,
   TooltipProvider,
