+++ packages/design-system/src/scalars/docs/getting-started.mdx
@@ -1,22 +1,20 @@
 import { Meta, Source } from '@storybook/blocks';
 
-<Meta title="Document Engineering/Docs/Getting started" />
+<Meta title="Document Engineering/Getting started" />
 
 # Installation
 
 <Source language="bash" code={`
-npm install --save @powerhousedao/design-system react-hook-form
+npm install --save @powerhousedao/design-system
 `}/>
 
 <Source language="bash" code={`
-yarn add @powerhousedao/design-system react-hook-form
+yarn add @powerhousedao/design-system
 `}/>
 
 <Source language="bash" code={`
-pnpm add @powerhousedao/design-system react-hook-form
+pnpm add @powerhousedao/design-system
 `}/>
-
-> Note: [React Hook Form](https://www.react-hook-form.com/) is a peer dependency to use built-in scalars validation.
 
 # Built-in scalars
