--- packages/design-system/src/ui/components/data-entry/index.ts
@@ -1,9 +1,0 @@
-export {
-  Checkbox,
-  type CheckboxProps,
-  type CheckboxValue,
-} from "./checkbox/index.js";
-export { Input, type InputProps } from "./input/index.js";
-export { TextInput, type TextInputProps } from "./text-input/index.js";
-export { Textarea, type TextareaProps } from "./textarea/index.js";
-export { Toggle, type ToggleProps } from "./toggle/index.js";
